feat: support rotating credentials in newsigv4
This adds an additional feature set to a fork of the sigv4
extension. An explicit configuration option can define
a shared credential file location and profile name, this
file will be watched for changes and propogate the fresh
credentials to upstream consumers.

Where possible the AWS SDK machinery has been leveraged
to maintain best compatibility with other AWS toolchains.

The change should maintain backwards compatibility with
the existing sigv4 extension, and both should be
interchangable for uses requiring static credentials or
sts credentials.

Closes https://github.com/rocketsciencegg/aws-gamelift/issues/174