fix(newsigv4): persists the watch event beyond the first instance
Inotify works on watching specific inodes, the current
implementation proved brittle if the process mutating
the watched file does so by created a new file and then
renaming. This action causes the file under the same
path to change contents and exist with a new inode.
Instead we watch the parent directory and filter incoming
events by filename.